+++ changelog.md
@@ -11,6 +11,7 @@
 * `[audio]` AudioMux tests
 * `[audio]` Added vibration logic (effect is visible in logs only)
 * `[settings]` Added DatabaseAgent base class and SettingsAgent.
+* `[audio]` Added vibration/enable sound database logic
 
 
 ### Changed